:recycle: use an interface to define PDF rasterization
